+++ core/mixins/index.js
@@ -11,3 +11,4 @@ export * from './StableIdSupport';
 export * from './XhIdSupport';
 export * from './ManagedSupport';
 export * from './LoadSupport';
+export * from './LogSupport';